[delete][#32] 기존 파일 (listarea, listItemView) 삭제
